refactor: auction time remaining to hook
To allow logic reuse
